+++ app/Http/Controllers/UserController.php
@@ -5,9 +5,11 @@ namespace App\Http\Controllers;
 use App\Models\User;
 use App\Models\Client;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
 
 class UserController extends Controller
 {
@@ -18,8 +20,9 @@ class UserController extends Controller
      */
     public function index()
     {
+        $users = User::whereIn('role', ['client', 'programmer'])->get();
         return view('user.index', [
-            'users' => User::all()
+            'users' => $users
         ]);
     }
 
@@ -42,25 +45,43 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $validatedData = $request->validate([
+        $validator = Validator::make($request->all(), [
             'name' => 'required|max:255',
-            'email' => 'required|unique:users',
+            'email' => [
+                'required',
+                'email',
+                Rule::unique('users', 'email')->ignore($request->user),
+            ],
+            'email_verified_at' => 'required',
             'password' => 'required|min:5|max:255',
             'image' => 'image|file|mimes:jpg,png,jpeg,gif,svg',
             'role' => '',
-            'client_id' => ''
+            'client_id' => [
+                'required',
+                Rule::unique('users', 'client_id'),
+            ],
+        ], [
+            'email.unique' => 'Email Sudah Digunakan.',
+            'client_id.unique' => 'Perusahaan Klien Sudah Terdaftar.',
         ]);
+
+        if ($validator->fails()) {
+            return redirect()->back()
+                ->withErrors($validator)
+                ->withInput();
+        }
 
         $user = new User();
         $user->name = $request->name;
         $user->email = $request->email;
+        $user->email_verified_at = $request->email_verified_at;
         $user->password = Hash::make($request->password);
         $user->image = $request->file('image')->store('images');
         $user->role = $request->role;
         $user->client_id = $request->client_id;
         $user->save();
 
-        return redirect('/user')->with('success','Data Baru Berhasil Ditambahkan!');
+        return redirect('/user')->with('success', 'Data Baru Berhasil Ditambahkan!');
     }
 
     /**
@@ -83,8 +104,10 @@ class UserController extends Controller
     public function edit($id)
     {
         $user = User::find($id);
+        $client = Client::all();
         return view('user.edit', compact(
-            'user'
+            'user',
+            'client'
         ));
     }
 
@@ -99,21 +122,57 @@ class UserController extends Controller
     {
         $rules = [
             'name' => '',
-            'email' => '',
-            'image' => 'image|file|mimes:jpg,png,jpeg,gif,svg|max:2048'
+            // 'password' => 'nullable|min:5|max:255',
+            'image' => 'image|file|mimes:jpg,png,jpeg,gif,svg|max:2048',
+            'client_id' => 'required'
         ];
 
-        $validatedData = $request->validate($rules);
+        // Validasi jika email sudah ada
+        $emailRule = Rule::unique('users', 'email')->ignore($id);
+        $rules['email'] = ['email', $emailRule];
 
-        if($request->file('image')){
-            if($request->oldImage) {
+        $clientRule = Rule::unique('users', 'client_id')->ignore($id);
+        $rules['client_id'] = $clientRule;
+
+        $customMessages = [
+            'email.unique' => 'Email Sudah Digunakan.',
+            'client_id.unique' => 'Perusahaan Klien Sudah Terdaftar.',
+        ];
+
+        $validator = Validator::make($request->all(), $rules, $customMessages);
+
+        if ($validator->fails()) {
+            return redirect()->route('user.edit', $id)
+                ->withErrors($validator)
+                ->withInput();
+        }
+
+        // Pengambilan data user berdasarkan ID
+        $user = User::find($id);
+
+        if (!$user) {
+            return redirect()->route('user.edit', $id)
+                ->with('error', 'User tidak ditemukan!');
+        }
+
+        $validatedData = $validator->validated(); // Mengambil data yang telah divalidasi
+
+        // if ($validatedData['password']) {
+        //     // Jika password diisi pada form update, maka hash password baru menggunakan bcrypt
+        //     $validatedData['password'] = Hash::make($validatedData['password']);
+        // } else {
+        //     // Jika password tidak diisi pada form update, hapus key 'password' dari validatedData
+        //     unset($validatedData['password']);
+        // }
+
+        if ($request->file('image')) {
+            if ($request->oldImage) {
                 Storage::delete($request->oldImage);
             }
             $validatedData['image'] = $request->file('image')->store('images');
         }
 
-        User::where('id', $id)
-            ->update($validatedData);
+        $user->update($validatedData);
 
         return redirect('/user')->with('success', 'Data Berhasil Di Update!!');
     }
@@ -128,6 +187,6 @@ class UserController extends Controller
     {
         $user = User::findOrFail($id);
         $user->delete();
-        return redirect('/user')->with('success','Data Telah Berhasil Di Hapus!!');
+        return redirect('/user')->with('success', 'Data Telah Berhasil Di Hapus!!');
     }
 }
